Add feature to descend into subdir of parent dir
Instead of doing 'cd ../../../parent/subdir' or 'cd `bd parent`/subdir'
you can do '. bd parent/subdir' to get the same result.
The argument provided is split into a search string and a descend path.
The search string is used instead of $2 (or $1) in the main code, so
-s and -si work the same way as before,
and the descend path is appended to $NEWPWD if match is found.

Note: the argument is split at the first '/' only so its possible
to do something like '. bd parent/subdir/subdir2/subdir3'